Refactor the websocket notifications observer as concern
Having it as a concern instead of observer allows us to do more and this also means we only need to refactor the cache sweeper now to get rid of the activerecord-observers gem.

The main motivation for this was to also include the new tickets into the websocket notifications framework.