Fix websocket monitor script
Since we now require a logged in user to open a websocket connection we now need to first log in and then get the session cookie.

We now also require only the bare minimum of gems needed for the script and get the list of display channels to monitor from the new json api.