BUGFIX: Inkscape 0.91 has different behavior on empty lines
Previous inkscape version was ok having a row with empty/whitespace only tspans in it.

But the new 0.91 version throws a fit when it sees such a beast, the whole text element will be positioned at -1.967, 568.575 (don't ask why) and it will snap back to the said coordinates if dragged inside inkscape (don't ask why).

This brought up that we have a bug in our line layout code, where if the first line is just a newline we produce a such line with empty tspans, so juggling the if checks a little everything is fine again \o/