Partial bugfix: slide image links
Previously after the slide html panel gets updated by ajax the slide image would always link to the full sized slide image. Now we instead link to the slide details page.

Need to still fix it so that in the slide details page the link is always to the full sized slide image.